+++ src/main/java/com/Library/model/ComicBooks.java
@@ -17,7 +17,7 @@ public class ComicBooks {
 	@SequenceGenerator(name="sequence",sequenceName = "comicBooks_id_seq", allocationSize=1)
 	private int id;
 	private String nome;
-	private float nota;
+	private double nota;
 	
 	@OneToOne(mappedBy = "comicBooks")
 	private Photos photos;
@@ -38,11 +38,11 @@ public class ComicBooks {
 		this.nome = nome;
 	}
 	
-	public float getNota() {
+	public double getNota() {
 		return nota;
 	}
 	
-	public void setNota(float nota) {
+	public void setNota(double nota) {
 		this.nota = nota;
 	}
 
@@ -59,7 +59,7 @@ public class ComicBooks {
 		return "ComicBooks [id=" + id + ", nome=" + nome + ", nota=" + nota + ", photos=" + photos + "]";
 	}
 
-	public ComicBooks(int id, String nome, int nota, Photos photos) {
+	public ComicBooks(int id, String nome, double nota, Photos photos) {
 		super();
 		this.id = id;
 		this.nome = nome;
